[issue-1] move uniprocess algorithm out of top-level engine
